tilmanx/requirements.txt
asgiref==3.5.0
attrs==21.4.0
dj-database-url==0.5.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
gunicorn==20.1.0
importlib-metadata==4.11.3
inflection==0.5.1
jsonschema==4.4.0
Markdown==3.3.6
names==0.3.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
zipp==3.7.0
